finish remaning renames
