+++ packages/core/src/common/errors.ts
@@ -36,7 +36,7 @@ export const HOTKEYS_WARN_DECORATOR_NO_METHOD = ns + ` @HotkeysTarget-decorated 
 export const HOTKEYS_WARN_DECORATOR_NEEDS_REACT_ELEMENT =
     ns + ` "@HotkeysTarget-decorated components must return a single JSX.Element or an empty render.`;
 
-export const HOTKEYS_TARGET2_CHILDREN_LOCAL_HOTKEYS =
+export const HOTKEYS_TARGET_CHILDREN_LOCAL_HOTKEYS =
     ns +
     ` <HotkeysTarget> was configured with local hotkeys, but you did not use the generated event handlers to bind their event handlers. Try using a render function as the child of this component.`;
